requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email-validator==2.1.0.post1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pillow==10.4.0
python-dotenv==1.0.0
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.8
simple-websocket==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
visitor==0.1.3
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.1.2
